Test division more precisely but also more efficiently.
The number is guaranteed to be exact so long as the denominator of the
fraction reduced to lowest terms is 2^n * 5 ^ m.

Limiting the exponents in the calculation of the denominator means that
there are fewer possible remainders and makes termination faster.

Slow termination should only happen in pathological cases, but with enormous
numbers in the denominator it is possible.

Signed-off-by: mulhern <amulhern@redhat.com>